Backend/requirements.txt
fastapi
uvicorn
pydantic
boto3
python-dotenv
SQLAlchemy
SQLAlchemy-Utils
psycopg2
